Clarify redaction behavior in comment
Updated the comment above the `_logger.UserLoggedIn` method call to specify that the `RedactedData` property of the `InnerData` property will be redacted if the `Transitive` property is set to `true`. Previously, the comment only mentioned that the `Name` and `Email` properties of the `User` object would be redacted, and that the `RedactedData` property would not be redacted.